Update num-test-cases to be compatible with multiple LLMs and num_queries_per_prompt
Summary: The variable `num_test_cases` is defined to specify the number of prompts that need to be evaluated. However, if we are testing multiple LLMs or if we are conducting multiple rounds of testing queries per prompt, these factors will not be reflected in num_test_cases during the expansion/judge step. This discrepancy can lead to a mismatch between the number of responses and the final result count.

Reviewed By: YueLi28

Differential Revision: D56826808

fbshipit-source-id: b1fde2cca5d62788fa0ed3ab8df4ac5965a1be55