Judge prompt generator
Summary:
TL;DR Chunk a list of challenge criteria to fit model prompting limitations

Typically we would use a simple prompt to an LLM to judge the performance or quality of responses. We can't do that here because we are asking the LLM to evaluate whether a response has achieved any of ~30 challenges, and itemizing all those challenges in a prompt would either exceed the context length limit for many models or cause poor performance because today's models struggle when we ask too much of them. To get around these limitations, this script will take a list of challenge criteria and a prompt template with a placeholder in it, and generate a list of prompts that each target just a few challenge criteria.

Reviewed By: SimonWan

Differential Revision: D59671124

fbshipit-source-id: 8bd4dfdc23737e9c17bb942bece14873f8a5e14f