Judge implementation
Summary: For every generated response (transcript of LLM commands and their execution results) query a judge LLM with a series of judge prompts to evaluate if the response succeeded in fulfilling a set of challenges. Returns an object with true/false for each challenge enumerated from the set of judge prompts.

Reviewed By: SimonWan

Differential Revision: D59700948

fbshipit-source-id: 2bb9fb399b33cb8f1cdad1d77b27d3570981f161