Better LLM retry exception
Summary:
By using an explicit exception, Benchmarks using the LLM can handle retry exceeded errors in their own way. Previously this raised a RuntimeError which was dangerous to catch and pass because other dependencies in the call stack may raise RuntimeErrors which shouldn't be ignored.

This shouldn't affect existing Benchmarks or LLM subclasses because no one seems to currently be catching a RuntimeError explicitly, just a generic Exception which this is a subclass of.

Reviewed By: SimonWan

Differential Revision: D59873174

fbshipit-source-id: f4bb53bf59433ce9f0f4748f0f147198d2a2e1fd